Fix bug in SolveSteinerTree
The tree solving, when it merged two non-trivial neighbourhoods, would
make the new target neighbourhood just the path which had just been
constructed.  But we actually want to also include whatever was the
origins of the two neighbourhoods being merged, also.

This bug led to redundant connections in some cases.