Filter pipe choices more carefully
Some mods (e.g. Advanced Fluid Handling) add a bunch of weird sorts of
pipes which don't have their connections laid out in the way this mod
expects.  In that case, this mod will do the wrong thing in the layout
algorithm.

To avoid that, filter out pipe prototypes which don't match our
expectations.  This applies to both the regular and underground pipes.