Flood fill to prevent certain failures
It's possible for the pipe or heat pipe layout to fail because there's a
small isolated part of the map that cannot connect with the rest.

I've seen this happen for Vulcanus vents, which are quite closely
packed.

Prevent this from being a problem by adding a new step to
SolveSteinerTree which first finds all connected components of the
graph, and picks just the largest connected component to search for a
spanning tree in.