Handle invalid entities in ForceGhostAt
In the loop where we are marking entities for deconstruction, when in
editor mode, marking a cliff for deconstruction instantly removes it.
That in turn can cause connected cliff segments to cease to exist.
Which can make later entities in this loop invalid.

This could cause a crash.  Fix by checking for entity validity before
marking for deconstruction.